print half of the string while skipping 1

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -7,7 +7,7 @@
  */
 void puts2(char *str)
 {
-	int index;
+	unsigned long int index;
 	char letter;
 
 	for (index = 0; index < strlen(str); index + 2)
